+++ benchmarks/ObserverOverheadBenchmark.cs
@@ -2,8 +2,8 @@ using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
 
 using Veggerby.Boards;
-using Veggerby.Boards.Chess;
 using Veggerby.Boards.Artifacts.Relations;
+using Veggerby.Boards.Chess;
 using Veggerby.Boards.Flows.Events;
 using Veggerby.Boards.Flows.Observers;
 using Veggerby.Boards.Flows.Phases;